backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
pdfplumber==0.11.4
requests==2.32.3
pytz==2024.2
sqlparse==0.5.1
